Search model conformance to Equatable / Hashable.
