+++ NextRace/NextRace/Sources/Views/Tabs/Search/SearchScreen/SearchViewModel.swift
@@ -63,30 +63,6 @@ final class SearchViewModel: ObservableObject {
     }
 }
 
-struct SearchResult: Equatable {
-    let page: Page
-    let races: [Race]
-}
-
-struct Race: Equatable, Identifiable {
-    let id: String
-    let name: String
-    let imageURL: String
-    let venue: Venue
-    let date: Date?
-    let seatmapURL: String?
-    let price: PriceRanges?
-
-    struct Venue: Equatable {
-        let name: String
-        let postalCode: String
-        let city: String
-        let state: String
-        let country: String
-        let address: String
-    }
-}
-
 enum Championship: String, CaseIterable, Identifiable {
     case formula
     case monster
